@@ -1,193 +1,126 @@
 plugins {
-    id 'org.springframework.boot' version '2.1.10.RELEASE'
-    id 'io.spring.dependency-management' version '1.0.8.RELEASE'
+    id 'org.springframework.boot' version '2.7.1'
+    id 'io.spring.dependency-management' version '1.0.11.RELEASE'
     id 'java'
-    id "com.bmuschko.docker-remote-api" version "4.4.1"
-    id "org.unbroken-dome.test-sets" version "2.1.1"    
 }
 
+apply plugin: 'io.spring.dependency-management'
 apply from: "$rootDir/gradle/versioning.gradle"
 
 group = 'ch.so.agi.ilivalidator'
 
 description = """ilivalidator web service"""
 
-// A launch script will be added. Therefor it is
-// not a jar anymore but zip file and jar xf *.jar
-// will not work.
-/*
+sourceCompatibility = '17'
+targetCompatibility = '17'
+
+
+compileJava {
+    options.compilerArgs.addAll(['--release', '11'])
+}
+
+
 bootJar {
-    launchScript()
-}
-*/
-
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
-
-tasks.withType(JavaCompile) {
-    options.encoding = 'UTF-8'
-}
-
-testSets {
-    dockerTest
+   archiveName = baseName + "-" + version + "-exec.jar"
 }
 
 sourceSets {
-    dockerTestRuntime {
-        runtimeClasspath += sourceSets.test.runtimeClasspath
+    dockerTest {
+        java.srcDir "$projectDir/src/dockerTest/java"
+        compileClasspath += main.output + test.output
+        runtimeClasspath += main.output + test.output
     }
+}
+
+configurations {    
+    dockerTestImplementation.extendsFrom testImplementation
+    dockerTestRuntime.extendsFrom testRuntime    
 }
 
 repositories {
-    jcenter()
-    maven { url "http://jars.interlis.ch/" }
-    maven { url "http://repo.maven.apache.org/maven2" }
+    mavenCentral()
+    maven { url "https://jars.interlis.ch/" }
 }
 
-configurations {
-    customFunctionsDeps
-    customFunctionsDeps.transitive = false
+configurations.all {
+    resolutionStrategy {
+        force 'ch.interlis:ili2c-tool:5.2.3'
+        force 'ch.interlis:ili2c-core:5.2.3'    
+    }
 }
 
 dependencies {
-    compile "org.springframework.boot:spring-boot-starter-web"
-    compile "org.springframework.boot:spring-boot-starter-thymeleaf"
-    compile "org.springframework.boot:spring-boot-devtools"
-    compile "org.springframework.boot:spring-boot-starter-actuator"
-    compile "commons-io:commons-io:2.5"
-    compile group: 'ch.interlis', name: 'ilivalidator', version: '1.11.3'
-    compile "org.webjars:materializecss:0.100.1"
-    compile "org.webjars:jquery:2.2.4"
-    compile "org.jgrapht:jgrapht-core:1.3.1" // for custom functions
-    compile 'ch.interlis:ili2db:4.3.2' // for custom functions
-    
-    customFunctionsDeps "ch.so.agi.ilivalidator:ilivalidator-custom-functions:1.0.17"
-    testCompile "io.rest-assured:rest-assured:3.0.6" 
-    testCompile("org.springframework.boot:spring-boot-starter-test") {
-        exclude(module: 'commons-logging')
+    implementation 'org.springframework.boot:spring-boot-starter-actuator'
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter-websocket'
+    implementation 'commons-io:commons-io:2.6'  
+    implementation 'ch.interlis:ilivalidator:1.11.15-SNAPSHOT'
+    implementation 'io.github.sogis:ilivalidator-custom-functions:1.1.47'
+    implementation platform('software.amazon.awssdk:bom:2.15.33')
+    implementation 'software.amazon.awssdk:s3'
+
+    developmentOnly 'org.springframework.boot:spring-boot-devtools'
+    testImplementation('org.springframework.boot:spring-boot-starter-test') {
+        exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
     }
-    dockerTestCompile "io.rest-assured:rest-assured:3.0.6" 
-    dockerTestCompile sourceSets.test.output      
+    testImplementation 'org.jsoup:jsoup:1.12.1'
+    testImplementation 'org.testcontainers:testcontainers:1.17.3'
+    testImplementation 'org.testcontainers:junit-jupiter:1.17.3' 
+    dockerTestImplementation sourceSets.test.output          
 }
 
-test {
+tasks.named('test') {
+    useJUnitPlatform()
     failFast = true
     testLogging {
         showStandardStreams = true
-    }
+    }   
 }
 
-task deleteFilesLibsExt(type: Delete) {
-    delete fileTree('src/main/resources/libs-ext/') {
-        include '**.jar'
-    }
+//def DOCKER_ACCOUNT = "sogis"
+def DOCKER_ACCOUNT = "edigonzales"
+def DOCKER_REPO = "ilivalidator-web-service"
+
+// only for testing
+task buildImage(type: Exec) {
+    workingDir "$projectDir"
+    commandLine 'docker', 'build',
+                '--no-cache', '--force-rm',
+                '-t', "$DOCKER_ACCOUNT/$DOCKER_REPO:latest",
+                '-f', 'Dockerfile',  '.'
 }
 
-task copyToLibsExt(type: Copy) {
-    from configurations.customFunctionsDeps
-    into "src/main/resources/libs-ext/"
-}
-copyToLibsExt.dependsOn(deleteFilesLibsExt)
-build.dependsOn(copyToLibsExt)
-classes.dependsOn(copyToLibsExt)
-
-task deleteFilesJarDir(type: Delete) {
-    delete fileTree("${buildDir}/dependency/")//.listFiles()
-    delete file("${buildDir}/dependency/")
-}
-
-task unzipJar(type: Copy) {
-    from zipTree(file("${buildDir}/libs/ilivalidator-web-service-"+version+".jar"))
-    into file("${buildDir}/dependency/")
-}
-unzipJar.dependsOn(deleteFilesJarDir)
-bootJar.finalizedBy(unzipJar)
-
-import com.bmuschko.gradle.docker.tasks.image.Dockerfile
-import com.bmuschko.gradle.docker.tasks.image.DockerBuildImage
-import com.bmuschko.gradle.docker.tasks.image.DockerPushImage
-import com.bmuschko.gradle.docker.tasks.container.DockerCreateContainer
-import com.bmuschko.gradle.docker.tasks.container.DockerStartContainer
-import com.bmuschko.gradle.docker.tasks.container.DockerLogsContainer
-import com.bmuschko.gradle.docker.tasks.container.DockerStopContainer
-import com.bmuschko.gradle.docker.tasks.container.DockerCreateContainer.ExposedPort
-
-docker {
-    registryCredentials {
-        username = System.env.dockerRegistryUser
-        password = System.env.dockerRegistryPass
-    }
-}
-
-task buildDockerImage(type: DockerBuildImage) {
-    description = "Build docker image."
-    inputDir = project.rootDir
-    //tag = "sogis/ilivalidator-web-service:${version}"
-    tags = ["sogis/ilivalidator-web-service:"+version, "sogis/ilivalidator-web-service:latest"]
-    labels = ["ilivalidator-web-service.created": new SimpleDateFormat("yyyy-MM-dd HH:mm:ss").format(new Date()), 
-            "ilivalidator-web-service.git_commit": getCheckedOutGitCommitHash(), 
-            "ilivalidator-web-service.build": version]
-}
-
-task createDockerContainer(dependsOn: 'buildDockerImage', type: DockerCreateContainer) {
-    description = "Create docker container."
-    targetImageId buildDockerImage.getImageId() 
-    portBindings = ['8888:8888']
-    exposedPorts = [new ExposedPort("tcp", [8888])]
+task dockerTest(type: Test) {
+    description = 'Runs the docker tests.'
+    group = 'verification'
     
+    useJUnitPlatform()
+    testClassesDirs = sourceSets.dockerTest.output.classesDirs
+    classpath = sourceSets.dockerTest.runtimeClasspath    
+    outputs.upToDateWhen { false }
+    mustRunAfter test
 }
 
-task startDockerContainer(dependsOn: 'createDockerContainer', type: DockerStartContainer) {	
-    description = "Start docker container."
-    targetContainerId createDockerContainer.getContainerId() 
+// When used local you have to create an buildx env: 'docker buildx create --use'
+// This is not necessary with Github Action.
+task buildAndPushImages(type: Exec) {
+    def githash = getCheckedOutGitCommitHash()
+    def buildTimestamp = getTimestamp()
 
-    // Starting container != web service is ready.
-    doLast { 
-        sleep(10*1000)
-    }
-}
-
-task logDockerContainer(type: DockerLogsContainer) {
-    description = "Show docker container logs on standard out."
-    targetContainerId startDockerContainer.getContainerId()
-    follow = true
-    tailAll = true
-}
-
-task stopDockerContainer(type: DockerStopContainer) {
-    description = "Stop running docker container."
-    targetContainerId createDockerContainer.getContainerId()
-}
-
-def tags = ["latest", version]
-
-tags.each { tag ->
-    def imageTag = tag.toString()
-    task "pushDockerImage_$imageTag"(type: DockerPushImage) {
-        description = "Push docker image to hub.docker.com: $imageTag"
-        imageName =  "sogis/ilivalidator-web-service:$imageTag"
-    }
-}
-
-task pushDockerImages() {
-    description = "Push docker images to hub.docker.com."
-    dependsOn {
-        tasks.findAll { task -> task.name.startsWith('pushDockerImage_') }
-    }    
-}
-
-
-// TODO: How to show docker log. Seems to stall b/c of finalizedBy?
-// At least standard out is in gradle docker test reports.
-//startDockerContainer.finalizedBy('logDockerContainer')
-
-dockerTest.dependsOn('startDockerContainer')
-dockerTest.finalizedBy('stopDockerContainer')//, 'pushDockerImage_One', 'pushDockerImage_Two')
-
-import java.text.SimpleDateFormat
-
-def getCheckedOutGitCommitHash() {
-    'git log -1 --pretty=%H'.execute().text.trim()
+    workingDir "$projectDir"
+    commandLine 'docker', 'buildx', 'build',
+                '--platform', 'linux/arm64,linux/amd64',
+                '--no-cache', '--force-rm', '--push',
+                '--label', "org.opencontainers.image.title=$project.description",
+                '--label', "org.opencontainers.image.version=$version",
+                '--label', "org.opencontainers.image.base.name=docker.io/bellsoft/liberica-openjdk-alpine-musl:17.0.3",
+                '--label', "org.opencontainers.image.created=$buildTimestamp", 
+                '--label', "org.opencontainers.image.revision=$githash",
+                '-t', "$DOCKER_ACCOUNT/$DOCKER_REPO:latest",
+                '-t', "$DOCKER_ACCOUNT/$DOCKER_REPO:$version.major",
+                '-t', "$DOCKER_ACCOUNT/$DOCKER_REPO:"+version.major+"."+version.minor,
+                '-t', "$DOCKER_ACCOUNT/$DOCKER_REPO:$version",
+                '-f', 'Dockerfile',  '.'
 }
 
 task versionTxt()  {
